Removed property and logic to list engines enum from class. Logic is now in the ASTClient property. Separated license classes per file.
